+++ mtg/src/main/scala/com/github/gchudnov/mtg/diagram/Renderer.scala
@@ -2,7 +2,7 @@ package com.github.gchudnov.mtg.diagram
 
 import com.github.gchudnov.mtg.Diagram
 import com.github.gchudnov.mtg.Diagram.Theme
-import com.github.gchudnov.mtg.diagram.internal.BasicRenderer
+import com.github.gchudnov.mtg.diagram.internal.AsciiRenderer
 
 /**
  * Renderer
@@ -16,4 +16,4 @@ object Renderer:
     make
 
   def make: Renderer =
-    new BasicRenderer()
+    new AsciiRenderer()
